Refactoring to move away from arrays
A Vec3 is a three-element value; not an array that is "likely" to have 3
elements in it (but maybe not)

+++ hitable_list.v
@@ -5,11 +5,11 @@ struct HitableList {
 fn (h HitableList) hit(r Ray, t_min f32, t_max f32, mut rec HitRecord) bool {
 	mut temp_rec := new_hit_record()
 	mut hit_anything := false
-	mut closest_so_far := f64(t_max)
+	mut closest_so_far := t_max
 	for i := 0; i < h.list.len; i++ {
-		if h.list[i].hit(r, t_min, f32(closest_so_far), mut temp_rec) {
+		if h.list[i].hit(r, t_min, closest_so_far, mut temp_rec) {
 			hit_anything = true
-			closest_so_far = f64(temp_rec.t)
+			closest_so_far = temp_rec.t
 			rec = temp_rec
 		}
 	}
